Added scala-js support
This is mainly switching to crossprojects in build.sbt and splitting of JVM specific tests.
