+++ project/plugins.sbt
@@ -3,3 +3,5 @@ addSbtPlugin("com.codecommit" % "sbt-spiewak-sonatype" % "0.19.2")
 addSbtPlugin("org.scalameta" % "sbt-mdoc" % "2.2.13")
 addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.4.2")
 addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.4.0")
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.3.1")
+addSbtPlugin("org.portable-scala" % "sbt-scalajs-crossproject" % "1.0.0")
